+++ platerplotter/awaitingsorting/views.py
@@ -6,8 +6,8 @@ from django.urls import reverse
 
 from holdingracks.forms import HoldingRackForm
 from holdingracks.holding_rack_manager import HoldingRackManager
-from holdingracks.models import HoldingRack, HoldingRackWell
-from notifications.models import Sample, ReceivingRack
+from platerplotter.models import HoldingRack, HoldingRackWell
+from platerplotter.models import Sample, ReceivingRack
 from problemsamples.forms import LogIssueForm, SampleSelectForm
 from problemsamples.views import confirm_sample_positions
