+++ platerplotter/holdingracks/forms.py
@@ -4,7 +4,7 @@ from django import forms
 from django.core.exceptions import ValidationError
 from django.forms import ModelForm
 
-from .models import Plate
+from platerplotter.models import Plate
 
 
 class HoldingRackForm(forms.Form):
