+++ platerplotter/problemsamples/forms.py
@@ -4,7 +4,7 @@ from django.core.exceptions import ValidationError
 from django.forms import ModelForm
 from django import forms
 
-from notifications.models import Sample
+from platerplotter.models import Sample
 
 
 class LogIssueForm(forms.ModelForm):
